feat(app): fractional scale support and pre-initialize app window
- Change PetConfig.scale from u32 to f32 to allow values like 1.5
- Update blit_frame to accept f32 scale and compute output size with rounding
- Pre-create AppWindowState at startup (always-alive) instead of re-creating on each open; track open/closed via app_window_open flag and app_window_gen for viewport ID uniqueness
- Pass pre-loaded SpriteGallery into AppWindowState to avoid double-load
- Add tests: test_float_scale (blit_scale_1_5, config roundtrip), test_window_state (AppWindowState API)

+++ src/config/schema.rs
@@ -20,8 +20,8 @@ pub struct PetConfig {
     pub state_machine: String,
     pub x: i32,
     pub y: i32,
-    /// Integer pixel-art upscale factor.
-    pub scale: u32,
+    /// Pixel-art upscale factor (fractional values supported, e.g. 1.5).
+    pub scale: f32,
     /// Pixels per second for walk.
     pub walk_speed: f32,
 }
@@ -34,7 +34,7 @@ impl Default for PetConfig {
             state_machine: "embedded://default".to_string(),
             x: 100,
             y: 800,
-            scale: 2,
+            scale: 2.0,
             walk_speed: 80.0,
         }
     }
